Add transmit command
This command is intended to demonstrate sending an invoice to SdI.

Added CA cert and verbose flags for better control from the command line.
The verbose flag provides more detailed debugging information about
the package's execution.
The ca-cert flag accepts a path to the CA Certificate.

Example usage of the command:

```console
$ go run ./cmd/gobl.fatturapa transmit --ca-cert ./ca-all.pem --cert ./SDI-B85905495.pem --key ./key_client.key --verbose invoice.xml
```

Additionally, separated responsibilities and added flexibility
in certificate storage.
The CLI handles the loading of the certificate path,
while the SdI package uses the loaded key.

Co-authored-by: Alex Malaszkiewicz <torrocus@users.noreply.github.com>
Co-authored-by: Grzeg Lisowski <noplisu@users.noreply.github.com>