Handle incoming SdI SOAP requests
Our SOAP server needs to be able to parse different types of incoming
requests. Here we prepare a structure that accepts the different types
of messages that SdI might send us and we parse the message.

To allow the application to handle the incoming requests the way it wants
(log/save the data) we make the handler function accept the a function that
gets the data from the request.

Co-authored-by: Alex Malaszkiewicz <torrocus@users.noreply.github.com>
Co-authored-by: Grzeg Lisowski <noplisu@users.noreply.github.com>